# back-end/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for app.",
        "title": "Viva",
        "contact": {},
        "version": "1.0"
    },
    "host": "3.14.29.134:8080",
    "paths": {
        "/api/all-constumer": {
            "get": {
                "description": "Retrieve Constumer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constumer"
                ],
                "summary": "Retrieve Constumer",
                "responses": {
                    "200": {
                        "description": "List of Constumer",
                        "schema": {
                            "$ref": "#/definitions/db.Customer"
                        }
                    }
                }
            }
        },
        "/api/all-item": {
            "get": {
                "description": "Retrieve Item Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mkt"
                ],
                "summary": "Retrieve Item Control",
                "responses": {
                    "200": {
                        "description": "List of Item Event",
                        "schema": {
                            "$ref": "#/definitions/db.Marketplace"
                        }
                    }
                }
            }
        },
        "/api/calc-cashback": {
            "post": {
                "description": "Calculation Cashback Total",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cashback"
                ],
                "summary": "Calculation Cashback Total",
                "parameters": [
                    {
                        "description": "For add new cashbacks in wallet to consumer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.Purchase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Used Create",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/constumer": {
            "post": {
                "description": "Create New Constumer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mkt"
                ],
                "summary": "Create New Constumer",
                "parameters": [
                    {
                        "description": "Data for create new Constumer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.CustomerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return for creat Constumer",
                        "schema": {
                            "$ref": "#/definitions/db.Customer"
                        }
                    }
                }
            }
        },
        "/api/constumer/{id}": {
            "get": {
                "description": "Retrieve an Constumer by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mkt"
                ],
                "summary": "Retrieve Constumer by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Constumer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Constumer Details",
                        "schema": {
                            "$ref": "#/definitions/db.Customer"
                        }
                    }
                }
            }
        },
        "/api/mkt": {
            "post": {
                "description": "Create New Item For marketPlace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mkt"
                ],
                "summary": "Create New Item For marketPlace",
                "parameters": [
                    {
                        "description": "Data for create new Item",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.MktInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return for creat item",
                        "schema": {
                            "$ref": "#/definitions/db.Marketplace"
                        }
                    }
                }
            }
        },
        "/api/mkt/{id}": {
            "get": {
                "description": "Retrieve an Item by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mkt"
                ],
                "summary": "Retrieve Item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item Details",
                        "schema": {
                            "$ref": "#/definitions/db.Marketplace"
                        }
                    }
                }
            }
        },
        "/api/modify-cashback": {
            "put": {
                "description": "Create new cashback",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Create cashback",
                "parameters": [
                    {
                        "description": "Data for modify the cashback ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.StorageCashBack"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cashback Create",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/modify-pushed": {
            "put": {
                "description": "Create new Pushed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Create Pushed",
                "parameters": [
                    {
                        "description": "Data for modify the pushed ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.StorageBuying"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pushed Create",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/modify-use": {
            "put": {
                "description": "Create new CashBacksUse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "CashBacksUse",
                "parameters": [
                    {
                        "description": "Data for modify the Used ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.StorageUse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Used Create",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/remove-cashback": {
            "post": {
                "description": "Remove Cashback Total",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cashback"
                ],
                "summary": "Remove Cashback Total",
                "parameters": [
                    {
                        "description": "For add new cashbacks in wallet to consumer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.Purchase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Used Create",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/storage/{id}": {
            "get": {
                "description": "Retrieve an storage by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Retrieve Storage by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Storage ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Storage Details",
                        "schema": {
                            "$ref": "#/definitions/db.Store"
                        }
                    }
                }
            }
        },
        "/api/storages": {
            "get": {
                "description": "Retrieve all storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Retrieve All Storage",
                "responses": {
                    "200": {
                        "description": "List of All Storage",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Store"
                            }
                        }
                    }
                }
            }
        },
        "/api/sync-customers": {
            "get": {
                "description": "Retrieve Constumer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constumer"
                ],
                "summary": "Sync Constumer for external api",
                "responses": {
                    "200": {
                        "description": "List of Constumer",
                        "schema": {
                            "$ref": "#/definitions/db.Customer"
                        }
                    }
                }
            }
        },
        "/api/test-token": {
            "post": {
                "description": "Cria um toke para auth do usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Test Token",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Token de autenticação",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token make:",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/token": {
            "get": {
                "description": "Cria um toke para auth do usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Criar token de auth",
                "responses": {
                    "200": {
                        "description": "token make:",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "xp": {
                    "type": "number"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
        "db.Marketplace": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "product": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "db.OpeningHours": {
            "type": "object",
            "properties": {
                "friday": {
                    "type": "string"
                },
                "monday": {
                    "type": "string"
                },
                "saturday": {
                    "type": "string"
                },
                "sunday": {
                    "type": "string"
                },
                "thursday": {
                    "type": "string"
                },
                "tuesday": {
                    "type": "string"
                },
                "wednesday": {
                    "type": "string"
                }
            }
        },
        "db.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "buying": {
                    "type": "integer"
                },
                "cashBacksUse": {
                    "type": "integer"
                },
                "cashback": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "openingHours": {
                    "$ref": "#/definitions/db.OpeningHours"
                },
                "phone": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
        "interfaces.CustomerInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
        "interfaces.MktInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "product": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "interfaces.Purchase": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "interfaces.StorageBuying": {
            "type": "object",
            "properties": {
                "buying": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "interfaces.StorageCashBack": {
            "type": "object",
            "properties": {
                "cashback": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "interfaces.StorageUse": {
            "type": "object",
            "properties": {
                "buying": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}